next

+++ src/21-L1_Cache_Benchmark.cuh
@@ -60,10 +60,10 @@ void launchL1Benchmark(GpuInformation info, BenchmarkProperties prop, InfoPropDe
 
             float *hostTime = nullptr;
             cudaMallocHost((void **) &hostTime, (sizeof(float) * 32));
-            float *hostLoad = nullptr;
-            cudaMallocHost((void **) &hostLoad, (sizeof(float) * 1024));
-            unsigned int *deviceTime = nullptr;
-            cudaMalloc((void **) &deviceTime, (sizeof(unsigned int) * 32));
+            unsigned int *hostLoad = nullptr;
+            cudaMallocHost((void **) &hostLoad, (sizeof(unsigned int) * 1024));
+            float *deviceTime = nullptr;
+            cudaMalloc((void **) &deviceTime, (sizeof(float) * 32));
             unsigned int *deviceLoad = nullptr;
             cudaMalloc((void **) &deviceLoad, (sizeof(unsigned int) * 1024));
